finish task

+++ database/migrations/2022_03_27_110146_create_students_table.php
@@ -15,8 +15,9 @@ class CreateStudentsTable extends Migration
     {
         Schema::create('students', function (Blueprint $table) {
             $table->id();
+
             $table->string('name');
-            $table->bigInteger('order');
+            $table->unsignedInteger('order');
 
             $table->foreignId('school_id')->constrained()->cascadeOnDelete();
